requirements.txt
beautifulsoup4
lxml
requests
python-telegram-bot==20.7
